Adds BotCard component
Creates a reusable BotCard component to display bot information, including the bot's image and display name.
This change also includes styling for the BotCard.

Renames BotInfo component to BotCard.
Updates BotCategory component to import the new BotCard component.
