--- frontend/src/pageComponents/playBot/BotInfo.tsx
@@ -1,25 +1,0 @@
-import { BotInfo } from "../../constants/botsConfig";
-
-type BotCardProps = {
-    botInfo: BotInfo;
-}
-
-function BotCard({ botInfo }: BotCardProps) {
-    function getBotImage() {
-        if (botInfo?.botImage) {
-            return <img className="bot-image" src={botInfo.botImage} />
-        } else {
-            // TODO: Add default placeholder image
-            return null;
-        }
-    }
-
-    return (
-        <div className="bot-card-container">
-            {getBotImage()}
-            <p className="bot-name">{botInfo.botDisplayName}</p>
-        </div>
-    )
-}
-
-export default BotCard;
